WIP: Draft allTxs handling
This commit implements the recording all all transactions on ReqTx in
the state (allTxs) like in the spec. To make it work transaction pruning
needed to be done also differently (by applicability instead of by id)
and the tests indicate a quite substantial difference in behavior:

Before transactions would expire (TxExpired) when not applicable, now
they would still be picked up by a future ReqSn if they become
applicable.

This raises the question of how long we would keep transactions in the
allTxs? It could feel quite like "magic" and/or different to interacting
with the layer 1. Furthermore, re-applying transactions to seenUTxO on
each ReqSn is quite expensive (with thousands or millions of old txs
which would never get cleaned up!).