Trial 2: Callback multiple events and evolve chainState only locally
This yields a nice, generic interface for the callback. It allows us to
handle the situation where multiple observable transactions are in a
single block, but also keeps the interface to retrieve and provide the
chainState.

One question: Should we associate the newChainState with each event or
with the list of events? The latter would force the head logic layer to
deal with the chainState outside of 'update' (not per event level).