Trial 3: Just not write back the chain state in callback
This seems to also satisfy the current model tests, which indicates that
we are not (yet) exercising the case which led to this code originally.

That case was: Multiple commit transactions are in the same block and
each observation (callback) would get the empty Initializing state. So
we would overwrite resulting chain state from commits observed prior in
a block.